% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoILSA.R
\name{autoILSA}
\alias{autoILSA}
\title{Options for automatic functions}
\usage{
autoILSA(func = c("leaguetable", "proflevels"), study = NULL)
}
\arguments{
\item{func}{a character value indicating the options of which function should be printed.}

\item{study}{an optional character vector indicating the ILSA name.}
}
\value{
a list.
}
\description{
Shows all options for the automatic functions: \code{\link{leaguetable}}, and \code{\link{proflevels}}.
}
\examples{
autoILSA()

}
