% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaguetable.R
\name{leaguetable}
\alias{leaguetable}
\title{ILSA's league tables}
\usage{
leaguetable(
  df,
  study = NULL,
  year,
  subject = NULL,
  specification = NULL,
  addCI = TRUE,
  alpha = 0.05,
  method = NULL,
  reps = NULL,
  fixN = TRUE
)
}
\arguments{
\item{df}{a data frame.}

\item{study}{an optional character vector indicating the subjects to be analyzed, for a list of available
subjects, check \code{\link{ILSAinfo}}.}

\item{year}{a numeric vector indicating the ILSA name, for a list of available
cycles, check \code{\link{ILSAinfo}}.}

\item{subject}{an optional character vector indicating the subject for a list of available
ILSA, check \code{\link{ILSAinfo}}.}

\item{specification}{a character value indicating extra specification like grade
(e.g., \code{"G8"} for TIMSS) or subject (e.g., \code{"Math"} for TIMSSADVANCED).}

\item{addCI}{a logical value indicating if confidence intervals should be added.
Defaults is \code{TRUE}.}

\item{alpha}{a numeric value indicating confidence level.}

\item{method}{a string indicating the name of the replication method.
Available options are:
\code{"JK2-full"}, \code{"JK2-half"},
\code{"FAY-0.5"},
and \code{"JK2-half-1PV"}. \cr\cr
Additionally, ILSA names can be used, defaulting into:
\itemize{
\item \code{"TIMSS"},  \code{"PIRLS"}, or \code{"LANA"} for \code{"JK2-full"};
\item \code{"ICILS"}, \code{"ICCS"}, or \code{"CIVED"} for \code{"JK2-half"};
\item \code{"PISA"} or \code{"TALIS"} for \code{"FAY-0.5"};
\item and \code{"oldTIMSS"}, \code{"oldPIRLS"}, or \code{"RLII"} for \code{"JK2-half-1PV"}.
}
Note that \code{"oldTIMSS"} and \code{"oldPIRLS"}
refer to the method used for TIMSS and PIRLS before 2015,
where within imputation variance
is estimated using only 1 plausible value.}

\item{reps}{an integer indicating the number of replications to be created.
If \code{NULL} the maximum number of zones will be used.}

\item{fixN}{a logical value indicating if data should be "fixed" to meet official criteria.
For example, reducing the sample for certain countries in TIMSS 1995. Default is \code{TRUE}.}
}
\value{
a data frame.
}
\description{
Estimates the mean score for all countries within a cycle of an ILSA.
Arguments \code{method}, \code{reps}, and \code{var}, are extracted from
\code{\link{ILSAinfo}} and can be overridden by the user.
}
\examples{
data(timss99)
leaguetable(df = timss99, year = 1999)

}
