% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{myfunc}
\alias{myfunc}
\title{My Wrapper Function}
\usage{
myfunc(x, method = formals(repse)$method, setup = formals(repse)$setup, ...)
}
\arguments{
\item{x}{Main input parameter}

\item{method, setup}{Parameters passed to [repse()]. Defaults:
- method: `r eval(formals(repse)$method)`
- setup: `r eval(formals(repse)$setup)`}

\item{...}{Additional arguments passed to [repse()]}
}
\value{
Description of return value
}
\description{
My Wrapper Function
}
