% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepILSA.R
\name{prepILSA}
\alias{prepILSA}
\title{Prepare ILSA Data}
\usage{
prepILSA(
  df,
  study = NULL,
  year = NULL,
  specification = NULL,
  fixN = TRUE,
  columns = NULL
)
}
\arguments{
\item{df}{a data frame.}

\item{study}{an optional character vector indicating the subjects to be analyzed, for a list of available
subjects, check \code{\link{ILSAinfo}}.}

\item{year}{a numeric vector indicating the ILSA name, for a list of available
cycles, check \code{\link{ILSAinfo}}.}

\item{specification}{a character value indicating extra specification like grade
(e.g., \code{"G8"} for TIMSS) or subject (e.g., \code{"Math"} for TIMSSADVANCED).}

\item{fixN}{a logical value indicating if data should be "fixed" to meet official criteria.
For example, reducing the sample for certain countries in TIMSS 1995. Default is \code{TRUE}.}

\item{columns}{a character vector indicating which columns should be selected.
If \code{NULL}, all columns will be selected.}
}
\description{
Modifies ILSA data to meet official participation cases, selects columns and
transforms data into simple data frames converting missing values to NAs.
}
\examples{

data(timss99)

head(timss99)

newdata <- prepILSA(df = timss99, columns = paste0("BSMMAT0",1:5),fixN = FALSE)

head(newdata)

}
