% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflevels.R
\name{proflevels}
\alias{proflevels}
\title{ILSA's proficiency levels}
\usage{
proflevels(
  df,
  study = NULL,
  year,
  subject = NULL,
  method = NULL,
  reps = NULL,
  type = c("long", "wide1", "wide2"),
  separateSE = TRUE,
  fixN = TRUE,
  accumulated = FALSE
)
}
\arguments{
\item{df}{a data frame.}

\item{study}{an optional character vector indicating the ILSA name, for a list of available
ILSA, check \code{\link{autoILSA}}. If \code{NULL}, the ILSA name will be determined
by the column names in the data frame.}

\item{year}{a numeric vector indicating the ILSA name, for a list of available
cycles, check \code{\link{autoILSA}}.}

\item{subject}{an optional character vector indicating the subject for a list of available
ILSA, check \code{\link{autoILSA}}.}

\item{method}{a string indicating the name of the replication method.
Available options are:
\code{"JK2-full"}, \code{"JK2-half"},
\code{"FAY-0.5"},
and \code{"JK2-half-1PV"}. \cr\cr
Additionally, ILSA names can be used, defaulting into:
\itemize{
\item \code{"TIMSS"},  \code{"PIRLS"}, or \code{"LANA"} for \code{"JK2-full"};
\item \code{"ICILS"}, \code{"ICCS"}, or \code{"CIVED"} for \code{"JK2-half"};
\item \code{"PISA"} or \code{"TALIS"} for \code{"FAY-0.5"};
\item and \code{"oldTIMSS"}, \code{"oldPIRLS"}, or \code{"RLII"} for \code{"JK2-half-1PV"}.
}
Note that \code{"oldTIMSS"} and \code{"oldPIRLS"}
refer to the method used for TIMSS and PIRLS before 2015,
where within imputation variance
is estimated using only 1 plausible value.}

\item{reps}{an integer indicating the number of replications to be created.
If \code{NULL} the maximum number of zones will be used.}

\item{type}{a character value indicating the type of table to produce.
Options include: \code{"long"}, for a long table with a column with the proportions
and another one for the standard error; \code{"wide1"} for a wide table where groups
are distributed in lines; \code{"wide2"} for a wide table where groups are distributed in columns.}

\item{separateSE}{a logical value indicating if standard errors should be separated from
proportions, each as an element from a list. Only works for wide tables. Default is \code{TRUE}.}

\item{fixN}{a logical value indicating if data should be "fixed" to meet official criteria.
For example, reducing the sample for certain countries in TIMSS 1995. Default is \code{TRUE}.}

\item{acumulated}{a logical value indicating if proficiency levels should be accumulated.}
}
\value{
a data frame or a list.
}
\description{
Estimates the proficiency levels for all countries within a cycle of an ILSA.
Arguments \code{method}, and \code{reps}, are extracted from
\code{\link{autoILSA}} and can be overridden by the user.
}
\examples{
data(timss99)

proflevels(timss99,year = 1999,type = "long",subject = "math")

}
