% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvse.R
\name{pvse}
\alias{pvse}
\title{Standard Error for Plausible Values}
\usage{
pvse(PVer, PVe0, df = FALSE)
}
\arguments{
\item{PVer}{to FILL}

\item{PVe0}{to FILL}

\item{df}{description}
}
\value{
the standard error and degrees of freedom.
}
\description{
Standard Error for Plausible Values
}
