% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaguetable.R
\name{leaguetable}
\alias{leaguetable}
\title{ILSA's league tables}
\usage{
leaguetable(
  df,
  study = NULL,
  year,
  subject = NULL,
  method = NULL,
  reps = NULL,
  var = c("unbiased", "ML")
)
}
\arguments{
\item{df}{a data frame.}

\item{study}{an optional character vector indicating the subjects to be analyzed, for a list of available
subjects, check \code{\link{ILSAinfo}}.}

\item{year}{a numeric vector indicating the ILSA name, for a list of available
cycles, check \code{\link{ILSAinfo}}.}

\item{subject}{an optional character vector indicating the subject for a list of available
ILSA, check \code{\link{ILSAinfo}}.}

\item{method}{a string indicating the name of the replication method.
Available options are:
\code{"JK2-full"}, \code{"JK2-half"},
\code{"FAY-0.5"},
and \code{"JK2-half-1PV"}. \cr\cr
Additionally, ILSA names can be used, defaulting into:
\itemize{
\item \code{"TIMSS"} or \code{"PIRLS"} for \code{"JK2-full"};
\item \code{"ICILS"}, \code{"ICCS"}, or \code{"CIVED"} for \code{"JK2-half"};
\item \code{"PISA"} or \code{"TALIS"} for \code{"FAY-0.5"};
\item and \code{"oldTIMSS"}, \code{"oldPIRLS"}, or \code{"RLII"} for \code{"JK2-half-1PV"}.
}
Note that \code{"oldTIMSS"} and \code{"oldPIRLS"}
refer to the method used for TIMSS and PIRLS before 2015,
where within imputation variance
is estimated using only 1 plausible value.}

\item{reps}{an integer indicating the number of replications to be created.
If \code{NULL} the maximum number of zones will be used.}

\item{var}{a string indicating the method to use for the variance:
\code{"unbiased"} calculates the unbiased estimate (n-1); \code{"ML"}
calculates the maximum likelihood estimate.}
}
\value{
a data frame.
}
\description{
Estimates the mean score for all countries within a cycle of an ILSA.
Arguments \code{method}, \code{reps}, and \code{var}, are extracted from
\code{\link{ILSAinfo}} and can be overridden by the user.
}
\examples{

leaguetable(df = timss99, year = 1999)

}
