% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflevels.get.R
\name{proflevels.get}
\alias{proflevels.get}
\title{Get ILSA's proficiency levels}
\usage{
proflevels.get(df, study = NULL, subject = NULL, combine = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{study}{an optional character vector indicating the subjects to be analyzed, for a list of available
subjects, check \code{\link{ILSAinfo}}.}

\item{subject}{an optional character vector indicating the subject for a list of available
ILSA, check \code{\link{ILSAinfo}}.}

\item{combine}{a logical value indicating if subjects should be combined in
a single data frame.}
}
\value{
a data frame or a list.
}
\description{
Converts ILSA scores into proficiency levels.
}
\examples{

proflevels.get(timss99, subject = "math")

}
